SaveChangeAsync Interceptor

+++ Core/App.Domain/Entities/Product.cs
@@ -11,7 +11,7 @@ namespace App.Domain.Entities
     {
         public string Name { get; set; }
         public int Stock { get; set; }
-        public long Price { get; set; }
+        public short Price { get; set; }
         public ICollection<Order> Orders { get; set; }
     }
 }
